Refactor Whats New modal to be part of React Navigation. Remove conflict between onboarding wizard and whats new modal
