+++ app/components/UI/DrawerView/index.js
@@ -53,7 +53,6 @@ import { newAssetTransaction } from '../../../actions/transaction';
 import { logOut, protectWalletModalVisible } from '../../../actions/user';
 import DeeplinkManager from '../../../core/DeeplinkManager';
 import SettingsNotification from '../SettingsNotification';
-import WhatsNewModal from '../WhatsNewModal';
 import InvalidCustomNetworkAlert from '../InvalidCustomNetworkAlert';
 import { RPC } from '../../../constants/network';
 import { findRouteNameFromNavigatorState } from '../../../util/general';
@@ -1170,7 +1169,6 @@ class DrawerView extends PureComponent {
 
     const {
       invalidCustomNetwork,
-      showProtectWalletModal,
       account: { name: nameFromState, ens: ensFromState },
       showModal,
       networkType,
@@ -1444,9 +1442,6 @@ class DrawerView extends PureComponent {
             showReceiveModal={this.showReceiveModal}
           />
         </Modal>
-        {showProtectWalletModal === false ? (
-          <WhatsNewModal navigation={this.props.navigation} enabled />
-        ) : null}
         {this.renderProtectModal()}
       </View>
     );
