+++ app/components/UI/WhatsNewModal/index.ts
@@ -1,0 +1,2 @@
+export { default } from './WhatsNewModal';
+export { whatsNew as whatsNewList } from './whatsNewList';
