+++ app/components/Views/Wallet/index.tsx
@@ -31,6 +31,9 @@ import OnboardingWizard from '../../UI/OnboardingWizard';
 import ErrorBoundary from '../ErrorBoundary';
 import { DrawerContext } from '../../Nav/Main/MainNavigator';
 import { useAppThemeFromContext, mockTheme } from '../../../util/theme';
+import { shouldShowWhatsNewModal } from '../../../util/onboarding';
+import Logger from '../../../util/Logger';
+import Routes from '../../../constants/navigation/Routes';
 
 const createStyles = (colors: any) =>
   StyleSheet.create({
@@ -124,6 +127,29 @@ const Wallet = ({ navigation }: any) => {
   const wizardStep = useSelector((state: any) => state.wizard.step);
 
   const { colors: themeColors } = useAppThemeFromContext() || mockTheme;
+
+  /**
+   * Check to see if we need to show What's New modal
+   */
+  useEffect(() => {
+    if (wizardStep > 0) {
+      // Do not check since it will conflict with the onboarding wizard
+      return;
+    }
+    const checkWhatsNewModal = async () => {
+      try {
+        const shouldShowWhatsNew = await shouldShowWhatsNewModal();
+        if (shouldShowWhatsNew) {
+          navigation.navigate(Routes.MODAL.ROOT_MODAL_FLOW, {
+            screen: Routes.MODAL.WHATS_NEW,
+          });
+        }
+      } catch (error) {
+        Logger.log(error, "Error while checking What's New modal!");
+      }
+    };
+    checkWhatsNewModal();
+  }, [wizardStep, navigation]);
 
   useEffect(
     () => {
