+++ app/constants/navigation/Routes.ts
@@ -17,6 +17,7 @@ const Routes = {
   MODAL: {
     DELETE_WALLET: 'DeleteWalletModal',
     ROOT_MODAL_FLOW: 'RootModalFlow',
+    WHATS_NEW: 'WhatsNewModal',
   },
   ONBOARDING: {
     ROOT_NAV: 'OnboardingRootNav',
